refactor: remove the causes of the extra compiler warnings
Give internal linkage to the functions only used in their own file,
include keyboard.h from keyboard.c, make the narrowing conversions of
the time code explicit, define Init after including confdefs.h, which
declares it, and drop the priority check that uint8_t makes always false.

+++ os/rtems5/src/periodic_timer.c
@@ -47,14 +47,14 @@ static rtems_interval get_sleep_time(const TimeVal * const next_time) {
 
         if (next_time->tv_usec >= (uint32_t)current_time.tv_usec) {
 
-            interval.tv_sec = next_time->tv_sec - current_time.tv_sec;
-            interval.tv_usec = next_time->tv_usec - current_time.tv_usec;
+            interval.tv_sec = (uint32_t)(next_time->tv_sec - current_time.tv_sec);
+            interval.tv_usec = next_time->tv_usec - (uint32_t)current_time.tv_usec;
 
         } else {
 
             interval.tv_sec = (uint32_t)(next_time->tv_sec - current_time.tv_sec) - 1;
             interval.tv_usec = 1000000 -
-                (uint32_t)(current_time.tv_usec - next_time->tv_usec);
+                ((uint32_t)current_time.tv_usec - next_time->tv_usec);
 
         }
 
@@ -77,7 +77,7 @@ static rtems_interval get_sleep_time(const TimeVal * const next_time) {
             TimeVal interval;
 
             interval.tv_sec = 0;
-            interval.tv_usec = next_time->tv_usec - current_time.tv_usec;
+            interval.tv_usec = next_time->tv_usec - (uint32_t)current_time.tv_usec;
 
             sleep_time = (rtems_interval)(interval.tv_usec / __TERMINA_MICROSECONDS_PER_TICK);
